file-splitter-project/python-splitter/requirements.txt
    fastapi==0.115.13
    uvicorn==0.34.3
    pydantic==2.7.1
    python-multipart==0.0.9
    aiofiles==23.2.1
    tqdm==4.66.4
    python-docx==1.1.0
    python-magic==0.4.27 # CHANGED THIS LINE
    